+++ pages/index.tsx
@@ -1,19 +1,16 @@
-import React, { useEffect, useState } from 'react'
-import MainLayout from '../layouts/Main'
+import React from 'react'
+import { MainLayout } from '../layouts/Main'
 
 // images
 import { PageMeta } from '../components/PageMeta'
 import styled from '@emotion/styled'
-import { InlineCode } from '../components/InlineCode'
-import { PrimaryLink } from '../components/PrimaryLink'
 import { SizedBox } from '../components/SizedBox'
 import { theme } from '../styles/theme'
-import { getPlatformName, OsPlatformName } from '../lib/utils'
-import { ExternalLink } from '../components/ExternalLink'
-import { useGithub } from '../lib/use-github'
 import { FeatureShowcase } from '../components/FeatureShowcase'
 import { Section } from '../components/Section'
 import { AutoplayVideo } from '../components/Video'
+import { PrimaryLink } from '../components/PrimaryLink'
+import Link from 'next/link'
 
 function Index() {
   return (
@@ -32,13 +29,16 @@ function Index() {
           blockchain
         </Title>
 
+        <Link href='/download'>
+          <PrimaryLink>Try Flowser desktop app</PrimaryLink>
+        </Link>
+
         <SizedBox height={theme.spacing.lg} />
 
-        <DownloadOptions />
-
-        <SizedBox height={theme.spacing.lg} />
-
-        <AutoplayVideo src='/videos/hero-demo.mp4' />
+        <AutoplayVideo
+          src='/videos/hero-demo.mp4'
+          style={{ width: '100%', maxHeight: '60vh' }}
+        />
       </LandingSection>
 
       <SizedBox height={theme.spacing.xl} />
@@ -71,89 +71,19 @@ function Index() {
   )
 }
 
-function DownloadOptions() {
-  const { latestRelease, winAsset, macArmAsset, macX64Asset } = useGithub()
-  const [platform, setPlatform] = useState<OsPlatformName>('MacOS')
-
-  useEffect(() => {
-    setPlatform(getPlatformName())
-  }, [])
-
-  if (platform === 'Linux') {
-    return (
-      <div>
-        <Notice>🚧 Linux is currently not yet supported. 🚧</Notice>
-        <Notice>
-          If you would benefit from adding Linux support, tell us here:{' '}
-          <ExternalLink href='https://github.com/onflowser/flowser/discussions/142'>
-            github.com/onflowser/flowser/discussions/142
-          </ExternalLink>
-        </Notice>
-      </div>
-    )
-  }
-
-  return (
-    <div>
-      <InstallMethod>
-        Run with Flow CLI
-        <SizedBox height={theme.spacing.sm} />
-        <InlineCode>flow flowser</InlineCode>
-      </InstallMethod>
-
-      <SizedBox height={theme.spacing.md} />
-      <OrText>or</OrText>
-      <SizedBox height={theme.spacing.md} />
-
-      <InstallMethod>
-        Download directly
-        <SizedBox height={theme.spacing.sm} />
-        {platform === 'Windows' ? (
-          <PrimaryLink href={winAsset?.browser_download_url} download>
-            Download for Windows
-          </PrimaryLink>
-        ) : (
-          <div style={{ display: 'flex' }}>
-            <PrimaryLink href={macArmAsset?.browser_download_url} download>
-              Download for Apple Silicon
-            </PrimaryLink>
-            <SizedBox width={10} />
-            <PrimaryLink href={macX64Asset?.browser_download_url} download>
-              Download for Apple Intel
-            </PrimaryLink>
-          </div>
-        )}
-        <SizedBox height={theme.spacing.md} />
-        <ReleaseLink href={latestRelease?.url}>
-          {latestRelease?.tagName ?? '-'}
-        </ReleaseLink>
-      </InstallMethod>
-    </div>
-  )
-}
-
 const LandingSection = styled.div`
-  max-width: 700px;
   margin: auto;
   text-align: center;
   display: flex;
   flex-direction: column;
+  align-items: center;
 `
 
 const Title = styled.h1`
   color: ${(props) => props.theme.color.light};
   font-size: ${(props) => props.theme.fontSize.big};
   font-weight: normal;
-`
-
-const Description = styled.p`
-  color: ${(props) => props.theme.color.grey};
-  font-size: ${(props) => props.theme.fontSize.md};
-  line-height: ${(props) => props.theme.lineHeight.lg};
-`
-
-const Notice = styled(Description)`
-  color: ${(props) => props.theme.color.light};
+  max-width: 700px;
 `
 
 const FlowHashTag = styled.span`
@@ -176,21 +106,6 @@ const FlowLogo = styled.img`
       transform: translateY(5px);
     }
   }
-`
-
-const InstallMethod = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  color: ${(props) => props.theme.color.light};
-`
-
-const OrText = styled.div`
-  color: ${(props) => props.theme.color.grey};
-`
-
-const ReleaseLink = styled(ExternalLink)`
-  color: ${(props) => props.theme.color.grey} !important;
 `
 
 export default Index
